xbps-uchroot: restrict arguments' position
GNU's getopt(3) and getopt_long(3) allows options anywhere in the
command. Thus, crippling (-O is used for cosmetic, but not required
to demonstrate the problem):

    xbps-uchroot -O /masterdir sh -c 'echo yes'

We could use:

    xbps-uchroot -O -- /masterdir sh -c 'echo yes'

to overcome that problem but it isn't documented in our man-pages.

Consider the limited usage of this utility, just stop parsing options
when we get our first non-options argument.

Signed-off-by: Doan Tran Cong Danh <congdanhqx@gmail.com>