Perform action with pkgs on hold only if -f is set.
As discussed with @duncaen in #273 we've decided it's better
to only update packages on hold mode when `-f` is set.

That means that `xbps-install -u foo` won't work anymore
with pkgs on hold unless `-f` is set.

This way one could perform a full system upgrade even with
pkgs on hold via `xbps-install -fu`.

cc @duncaen