lib/verifysig: use bearssl where possible.
BearSSL currently doesn't support decoding the pubkey in the format we
are using. Since there are plans for changing the signing algorithm,
this shouldn't be an issue for the transition, once the new design is
finalized. For now, we are using LibreSSL to decode the PEM string in
pubkey into a public key.