xbps-install: add --verify-sig option.
Forcing signature verification for local packages can be useful for
innumerous reasons, the simplest one being the possibility of adding a
test suite for this part of the code without requiring a test setup
running a server or similar.

For this, it was necessary to add a new flag value to xbps_handle, and I
took the opportunity to re-organize the code a bit, including always
checking sha256 for all packages and reporting when remove(3) fails.