lib/transaction_fetch: add xbps_file_sha256_check_raw.
This function allows us to check the sha256 sum for a particular file,
while at the same storing the calculated digest in the provided struct.

This allows us to always check the sha256 sum for packages during
installation, but also not have to scan the package twice when checking
its signature.

In order to simplify the code and necessary error checking, and because
some functions take a length parameter for the binary hash, while others
don't, introduce the xbps_sha256_digest struct, which guarantees we are
always passing a binary digest to functions which require one and
guarantees it has the necessary length to hold the digest, therefore
providing some form of type and memory safety and avoiding the need to
pass and propagate length parameters everywhere.