Create xbps_walk_dir function.
It's used in both xbps-create and xbps-uchroot. Even though neither
uses the "dirent" arg in fn, leave it in for possible future use.

This implementation also fixes a bug with the version of walk_dir in
those binaries where it leaks all the list[i] items. This version frees
them after using them and does a final cleanup in case the for loop
ended before going through all of them.

xbps-create requires this function to use alphasort, but xbps-uchroot
wouldn't. For implemenation simplicity, use alphasort in all cases.