tests/xbps/xbps-install: create tests for package conflicts
These tests are being added as a result of a real life bug: package B
conflicts with A, A starts depending on B (the conflict was solved, but
B wasn't updated to not conflict anymore), XBPS should not allow A to be
updated and install B as a dependency, but it does. Installing A
(instead of updating it) is blocked by XBPS with an error message about
the conflict between B has with A.

In this case it was an error in managing void-packages which caused the
situation, but XBPS's behavior should be consistent.

Initial commit causing issue:
https://github.com/void-linux/void-packages/commit/70eb5aec70ed8722515a318ae3e64b360d420741

Commit removing the conflict:
https://github.com/void-linux/void-packages/commit/fe505e9983bb6c19eb4ae1c7af54d45cb054a9bc