Add Meson
-Werror is dropped, because Meson's werror is more strict than configure
one.

This implementation does not support running tests. The Kyua testsuite
heavily relies on in-source builds. Tests would have to be rewritten
(at least the Kyuafile files, but probably more) to support Meson.

Meson does not support retrieving git info and providing it as a
compiler flag. Meson's vcs_tag() mechanism supports outputting to
a (header or source) file only. This would mean that xbps.h.in would
have to be modified.

No substitution is done on xbps.h.in. See #598 for explanation.

_DEFAULT_SOURCE is dropped, because _GNU_SOURCE implies _DEFAULT_SOURCE.

-D_FILE_OFFSET_BITS=64 is set automatically by Meson.

Work in progress!