lib: preserve: don't delete file's on package removal
The old behaviour was to only respect `preserve` on package updates.

This leads to issues where `vkpurge` would not be able to remove other
files not part of a kernel package when an obsolete kernel package was
removed using `xbps-remove -o`.

So for consistency `preserve` in a packages properties always means
`preserve`.