bin/xbps-create: add support for recording provides priority
Specified as a space-separated array of comma-separated keys and values
that record the priority which xbps should give to a virtual package a
package provides. Keys are pkgvers, and values are positive integers.
Higher values will be given a higher priority, and missing entries will
be given a priority of 0 by default.

This is implemented as a generic function with validation and parsing callbacks
so it can be used for other properties that are dictionaries.