bin/xbps-rindex: fix -c possibily merging the index and stage
when a package was removed while the repoistory was staged it would
write the merged index+stage into to the repository datas index
while also keeping the repository staged. This is fixed by operating
on the index without the stage. (maybe should have more clear variable
names than idx and index)