lib: handle replaces if replaced pkg is installed in the same transaction
replaces isn't honoured if the package to be replaced is currently being installed.
This fixes a corner case where a transitional package could be installed
along with its replacement, when it should be "ignored" and only the new
package installed.

This requires transaction commits to not fail if a replaced package
isn't installed when it should be "removed".

Also, indicate to the user that a package is being replaced, with
details about what replaced what shown in verbose output, not just debug output.

fixes: #667